Collision Detection
Created the logistics for collisions on platforms that determines the location of the player just before the collision. I however want to create a more efficient way of dealing with collisions with all platforms on the map that will prevent me from going through each platform individually. I will be working on that before moving forward to the next step.